requirements.txt
streamlit
openai
newsapi
smtplib
email
Pillow
replicate
json
dotenv
better_profanity







